Actualizaciones Varias

+++ ProyectoFinal_CristianLopez/Back-End/src/main/java/com/portfolio/CristianLopez/Security/Repository/UsuarioRepository.java
@@ -14,10 +14,11 @@ import org.springframework.stereotype.Repository;
  * @author krono
  */
 @Repository
-public interface UsuarioRepository extends JpaRepository<Usuario, Integer>{
+public interface UsuarioRepository extends JpaRepository<Usuario, Long>{
     Optional <Usuario> findByUsername(String username);
     
     boolean existsByUsername(String username);
     boolean existsByEmail(String email);
+    
 
 }
