+++ ProyectoFinal_CristianLopez/Back-End/src/main/java/com/portfolio/CristianLopez/service/ImpPersonaService.java
@@ -6,11 +6,13 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.portfolio.CristianLopez.repository.PersonaRepository;
+import java.util.Optional;
 
 @Service
 public class ImpPersonaService implements IPersonaService{
     
     @Autowired PersonaRepository ipersonaRepository;
+    
     
     @Override
     public List<Persona> verPersonas() {
@@ -20,17 +22,18 @@ public class ImpPersonaService implements IPersonaService{
 
     @Override
     public void guardarPersona(Persona pers) {
+      
         ipersonaRepository.save(pers);
     }
 
     @Override
-    public void borrarPersona(Long id) {
+    public void borrarPersona(Integer id) {
        ipersonaRepository.deleteById(id);
 
     }
 
     @Override
-    public Persona buscarPersona(Long id) {
+    public Persona buscarPersona(Integer id) {
         Persona pers = ipersonaRepository.findById(id).orElse(null); //orElse(null) permite enviar null si no encuentra la persona
         return pers;
     }
